+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Requires redis v4.0+ for multiple field/value pair support
+- Renamed pending key (TODO: need migration script
+
 ## [0.17.2] - 2021-06-06
 
 ### Fixed